Move the tests to the root `tests` folder.
