--- apps/integration-tests/game-simulation/src/app/GameSimulator.ts
@@ -1,6 +1,0 @@
-import { Beatmap } from "@osujs/core";
-import { Replay } from "node-osr";
-
-export interface GameSimulator {
-  simulate(beatmap: Beatmap, replay: Replay);
-}
