+++ tests/game-simulation/README.md
@@ -1,0 +1,7 @@
+Integration tests for `@osujs/(core|pp|math)` that require the testdata.
+
+You can run all testcases, but it takes a while since some beatmaps are just massive.
+
+```shell
+nx run tests-game-simulation:test
+```
