+++ tests/game-simulation/src/hitobjects.test.ts
@@ -1,10 +1,16 @@
 import { readFileSync } from "fs";
 import { Position } from "@osujs/math";
-import { getBlueprintFromTestDir } from "../src/app/util";
+import { getBlueprintFromTestDir } from "./util";
 import { buildBeatmap, Slider } from "@osujs/core";
 import { toMatchObjectCloseTo } from "jest-match-object-close-to";
 
 expect.extend({ toMatchObjectCloseTo });
+
+/**
+ * Tests the generation of the hitobjects against osu!lazer
+ *
+ * The test case files are generated with https://github.com/abstrakt8/osu
+ */
 
 interface Testsuite {
   filename: string;
@@ -19,19 +25,7 @@ interface Testsuite {
   }>;
 }
 
-
-// Before
-// 7182 / 9552 for precision = 9
-// 7125 / 9552 for precision = 11
-// 7124 / 9552 for precision = 12
-// 7124 / 9552 for precision = 12
-
-// Now
-// 9259 / 9552 for precision = 9
-
 const expectedPrecision = 4;
-// 384,307,21685,2,0,B|399:292|399:274|399:274|403:295|388:308,1,75.0000028610231,2|0,0:0|0:0,0:0:0:0:
-// 312,230,38626,2,0,P|306:266|323:298,1,75.0000028610231,2|0,1:2|0:0,0:0:0:0:
 
 function runTestSuite({ filename, sliders }: Testsuite) {
   describe(filename, function() {
